requirements.txt
requests
biopython
anthropic